Refactored Fascade API
I wanted to reduce the number of C files need to support the
python module, so I found a better way to implement the former
patch API. So added blynclight_api.[ch] and plastered over the
Embrava APIs with a new API that is simpler to me. The goal was
to reduce the number of functions exposed to the python module.