Rainbow less broke
The original implementation had a list of colors that was duplicated,
reversed and then added to the original list of colors. Turns out this
was a waste of space, only needed the original list.

The new API implementation resulted in weird blinking instead of a
gradual change in colors. Changed the way the sleep interval is
computed to slow it down a little.