Fixed fini_blynclights, added unique_device_id, bugfixes
The fini_blynclights signature changed from an int argument to
a void argument in the shared library. Updated the funcs array to
track the change.

Added unique_device_id method to track the function in the shared
object.

The shared object name changed from libblynclightcontrol.so to
libblynclight_api.so.

The BlyncLight.__init__ signature changed, updated the __repr__ to
track the reduced number of arguments.

Added unique_id method to BlyncLight and updated status method to
include unique_id.