refresh_blynclights replaced with sync_blynclights
The new sync_blynclights function takes an integer argument that
is the number of elapsed seconds to refresh since the last refresh.

Accompanying the change, the ndevices and last_update global varibles
in the shared object have been declared static. This means they can
not be directly accessed via ctypes and simplifies the interface.