Replaced BlyncLight_API with HID-based BlyncLight
The previous HID and ctypes based NewBlyncLight object was promoted to
BlyncLight. This facilitates the removal of the Embrava-based dynamic
library from the repository, which is a complication. The Human
Interface Device API provided by libhidapi, libhidapi-hidraw, or
libhidapi-libusb gives python direct access to USB devices via ctypes.
I've re-implemented pyhidapi as a submodule of blynclight to work
around a bug in MacOS. I've submitted a pull request to the
pyhidapi, but the project appears to have been abandonded.

The new BlyncLight implementation is a subclass of a ctypes structure
that defines a 9-byte command word that governs the behavior of the
light.