Added flash and speed properties
I was running into trouble setting flash and speed independently.
If on and flash == 1 and speed == 0, the light would turn off instead
of moving to a stead on state.  If flash was toggled from 0 to 1 with
speed == 0, again the light turns off.  So I renamed the flash and speed
bit fields to 'bflash' and 'bspeed' and wrapped then in properties
'flash' and 'speed' which coordinate the values of bflash and bspeed
so the light doesn't unexpectedly turn off.