BlyncLight object restructuring
The previous bytes property was broken and restructuring the light
object seemed like the easiest way to fix it. BlyncLight is now a
subclass of BlyncLightCommand which is a ctypes.Union of
BlyncLightBitfields and an array of 8 ctypes.c_ubyte. During the reorg
I found some bugs in the music bit fields which may be resolved.